LayoutTests/http/tests/privateClickMeasurement/attribution-conversion-through-cross-site-image-redirect-expected.txt

Tests that triggering of private click measurement conversions through cross-site redirects do not work.


Unconverted Private Click Measurements:
WebCore::PrivateClickMeasurement 1
Source: 127.0.0.1
Destination: localhost
Campaign ID: 3
No conversion data.
